LayoutTests/platform/chromium-android/media/media-can-play-mpeg-audio-expected.txt

Test HTMLMediaElement canPlayType() method with multiple audio mpeg MIME types.

These tests may be expected to fail if the WebKit port does not support the format.

EXPECTED (audio.canPlayType('audio/x-mpeg') == 'maybe'), OBSERVED '' FAIL
EXPECTED (audio.canPlayType('audio/x-mp3') == 'maybe') OK
EXPECTED (audio.canPlayType('audio/mp3') == 'maybe') OK
EXPECTED (audio.canPlayType('audio/mpeg') == 'maybe') OK
EXPECTED (audio.canPlayType('audio/mp4') == 'maybe') OK
EXPECTED (audio.canPlayType('audio/aac') == 'maybe'), OBSERVED '' FAIL
END OF TEST
